requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
psycopg2
django-jazzmin
drf-yasg
flake8
djangorestframework-simplejwt
django-filter
